Added some append and prepend functions and operators to Bu::List, now you can append one list to another and the like.
Also, wow, I found a bug that's been around for ages, I guess we don't copy
hash tables often.  The interesting thing is that it actually worked, it copied
but it would include any data that had been deleted in the old hash table but
not reclaimed yet and insert it as new data.  Usually the key had been
completely destroyed (like with a string) so it came out as keyed to blank
string.  So in cases like that, all deleted keys would collapse into one deleted
key in the new hash table.