Fixed a couple of things up, now build creates a shared object as well as a non shared, checkinst.sh makes sure that symlinks are in /usr/lib and /usr/include so that you don't need a libbu++ symlink in any directories.  If you still want the static version, then just delete the shared object, or switch your link line to do this:
-Wl,-Bstatic -lbu++ -Wl,-Bdynamic

Instead of just -lbu++, it'll include libbu++ as static, and everything else
as dynamic.  You could always just use -static, but then the whole thing is
going to be static.  Also, the dynamic library is already linked against all
the extra libs you need, so no need to link against pthread or libbz2.