***IMPORTANT*** The function Bu::Md5::getResult no longer returns a hex string, it returns the raw binary string that makes up the md5 sum, this matches the original goal of the API and makes the whole system more general and transportable.  I have added a handy helper function named getHexResult that will return the same classic hex md5 string we're used to, change anything in your code that uses getResult to getHexResult now.
I've also added a handy function to the CryptoHash to write the result to a
stream, writeResult.  I've fixed some more things in the formatter, and added
a cryptPass function that works very much like the system crypt function, md5
and base64.  If I knew more about the glibc implementation I could probably
make them compatible.  For now there are some subtle differences in the
formatting and the salting algorithm, also the output mantains it's base64
trailer (==) wheras the system function chops those off.  There's also another
helper that will only work on linux for now, that only takes the password, and
generates a salt for you using urandom.