Massive freaking changes!!!
Bu:;SharedCore actually is in and works, it's well tested and there are no
known memory leaks or violations as of now.  It's been applied to Bu::List and
Bu::FBasicString so far.  This means that everything using Bu::List and
Bu::FBasicString will be much, much faster and use considerably less memory.

I still have plans to apply this to Hash and maybe a couple of other core
classes.