Interesting tweak to the variant and optparser classes.  In the Variant, it would always fail if a const char * was passed in, it now converts these silently to Bu::FStrings, good to know...
Also, the OptParser now uses a Variant for overrides, meaning it doesn't have
to do extra parsing, and the amount of code you have to write may be
significantly reduced.  Pretty sweet, overall.  There is one downside.  For the
moment if you use a non-standard type or object as the target of a parameter
it always needs to have a formatter >> operator defined, even if you override
and the formatter >> operator is never called.  Hopefully we can get around
this in the future.

Also, it looks like it should be relatively trivial to create conversion
functions for the variant, they'll just be global template functions that
take two parameters, source type and target type.  Should be good times.